Move attributes query off model, onto view
This adds a `klass` argument, and renames a couple of attributes.
Other than that, this is just a move.

This function is gnarly. Fixing it can come later.
